functions/requirements.txt
# Add the Firebase SDK for Python
firebase-functions
firebase-admin
# Add the bit library for bitcoin transactions
bit
# Keep dotenv for local testing with .env files
python-dotenv
setuptools
# Add pycryptodome to provide ripemd160 hash
pycryptodome
